m2mresource: save RAM and ROM at M2MExecuteParameter
Change the API, use and internal storage of M2MExecuteParameter.
No need to allocate temp object from heap and copy the value just
for the duration of execute callback. This also reduces code
footprint by 120B on GCC.

Note: this changes the user facing API from String's to const char*
if the MEMORY_OPTIMIZED_API is set.

Also the get_argument_value() is now unconditionally returning a
const uint8_t pointer instead of uint8_t*, but that seems to me
a bug fix, as the callback client should not attempt to modify
the value given to it.